Move public test params to BaseStripeTest, add CustomerTest, and refactor ChargeTest
