+++ src/test/java/com/stripe/StripeTest.java
@@ -103,23 +103,23 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 public class StripeTest {
-	static Map<String, Object> defaultCardParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultSourceParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultDebitCardParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultChargeParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultCustomerParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultPlanParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultCouponParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultTokenParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultDebitTokenParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultBankAccountParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultRecipientParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultBitcoinReceiverParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultAlipayTokenParams = new HashMap<String, Object>();
-	static Map<String, Object> defaultManagedAccountParams = new HashMap<String, Object>();
-	static RequestOptions supportedRequestOptions;
+	public static Map<String, Object> defaultCardParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultSourceParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultDebitCardParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultChargeParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultCustomerParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultPlanParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultCouponParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultTokenParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultDebitTokenParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultBankAccountParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultRecipientParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultBitcoinReceiverParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultAlipayTokenParams = new HashMap<String, Object>();
+	public static Map<String, Object> defaultManagedAccountParams = new HashMap<String, Object>();
+	public static RequestOptions supportedRequestOptions;
 
-	static String getYear() {
+	public static String getYear() {
 		Date date = new Date(); //Get current date
 		Calendar calendar = new GregorianCalendar();
 		calendar.setTime(date);
@@ -226,16 +226,6 @@ public class StripeTest {
 		Map<String, Object> recipientParams = new HashMap<String, Object>();
 		recipientParams.putAll(defaultRecipientParams);
 		return Recipient.create(recipientParams);
-	}
-
-	static Customer createDefaultCustomerWithDefaultBitcoinReceiver()
-			throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams);
-		BitcoinReceiver receiver = BitcoinReceiver.create(defaultBitcoinReceiverParams);
-		Map<String, Object> createParams = new HashMap<String, Object>();
-		createParams.put("source", receiver.getId());
-		customer.getSources().create(createParams);
-		return customer;
 	}
 
 	static Map<String, Object> getSubscriptionParams() throws StripeException {
@@ -679,308 +669,6 @@ public class StripeTest {
 		// TODO: find a more reliable way to do this instead of sleeping
 		Thread.sleep(10000);
 		return Charge.retrieve(charge.getId());
-	}
-
-	// Customer Tests:
-	@Test
-	public void testCustomerCreate() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams, supportedRequestOptions);
-		assertEquals(customer.getDescription(), "J Bindings Customer");
-		List<Card> customerCards = customer.getCards().getData();
-		assertEquals(1, customerCards.size());
-		assertEquals("4242", customerCards.get(0).getLast4());
-	}
-
-	@Test
-	public void testCustomerCreateWithShippingDetails() throws StripeException {
-		ShippingDetails shippingDetails = new ShippingDetails();
-		shippingDetails.setName("name");
-		shippingDetails.setPhone("123-456-7890");
-		Address address = new Address()
-				.setCity("Washington")
-				.setCountry("USA")
-				.setLine1("1600 Pennsylvania Ave.")
-				.setLine2("line 2 address")
-				.setPostalCode("20500")
-				.setState("D.C.");
-		shippingDetails.setAddress(address);
-
-		Map<String, Object> params = ImmutableMap.<String, Object>builder()
-				.putAll(defaultCustomerParams)
-				.put("shipping", ImmutableMap.builder().put("address", ImmutableMap.builder().put("line1", address.getLine1()).put("line2", address.getLine2()).put("city", address.getCity()).put("country", address.getCountry()).put("postal_code", address.getPostalCode()).put("state", address.getState()).build()).put("name", shippingDetails.getName()).put("phone", shippingDetails.getPhone()).build())
-				.build();
-
-		Customer customer = Customer.create(params);
-		assertEquals(customer.getShipping(), shippingDetails);
-	}
-
-	@Test
-	public void testCustomerRetrieve() throws StripeException {
-		Customer createdCustomer = Customer.create(defaultCustomerParams);
-		Customer retrievedCustomer = Customer.retrieve(createdCustomer.getId());
-		assertEquals(createdCustomer.getCreated(),
-				retrievedCustomer.getCreated());
-		assertEquals(createdCustomer.getId(), retrievedCustomer.getId());
-	}
-
-	@Test
-	public void testCustomerList() throws StripeException {
-		Map<String, Object> listParams = new HashMap<String, Object>();
-		listParams.put("count", 1);
-		List<Customer> Customers = Customer.all(listParams).getData();
-		assertEquals(Customers.size(), 1);
-	}
-
-	@Test
-	public void testCustomerUpdate() throws StripeException {
-		Customer createdCustomer = Customer.create(defaultCustomerParams);
-		Map<String, Object> updateParams = new HashMap<String, Object>();
-		updateParams.put("description", "Updated Description");
-		Customer updatedCustomer = createdCustomer.update(updateParams);
-		assertEquals(updatedCustomer.getDescription(), "Updated Description");
-	}
-
-	@Test(expected=InvalidRequestException.class)
-	public void testCustomerUpdateToBlank() throws StripeException {
-		Customer createdCustomer = Customer.create(defaultCustomerParams);
-		Map<String, Object> updateParams = new HashMap<String, Object>();
-		updateParams.put("description", "");
-		createdCustomer.update(updateParams);
-	}
-
-	@Test
-	public void testCustomerUpdateToNull() throws StripeException {
-		Customer createdCustomer = Customer.create(defaultCustomerParams);
-		Map<String, Object> updateParams = new HashMap<String, Object>();
-		updateParams.put("description", null);
-		Customer updatedCustomer = createdCustomer.update(updateParams);
-		assertEquals(updatedCustomer.getDescription(), null);
-	}
-
-	@Test
-	public void testCustomerSourceList() throws StripeException {
-		Customer customer = createDefaultCustomerWithDefaultBitcoinReceiver();
-
-		HashMap<String, Object> listParams = new HashMap<String, Object>();
-		List<ExternalAccount> customerSourceList = customer.getSources().all(listParams).getData();
-
-		assertEquals(2, customerSourceList.size());
-		assert(customerSourceList.get(0) instanceof Card);
-		assertEquals("4242", ((Card) customerSourceList.get(0)).getLast4());
-		assert(customerSourceList.get(1) instanceof BitcoinReceiver);
-		assertEquals(true, ((BitcoinReceiver) customerSourceList.get(1)).getFilled());
-	}
-
-	@Test
-	public void testCustomerSourceRetrieveCard() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams);
-		ExternalAccountCollection customerSources = customer.getSources();
-		String paymentSourceId = customerSources.getData().get(0).getId();
-		ExternalAccount paymentSource = customerSources.retrieve(paymentSourceId);
-		assertNotNull(paymentSource);
-		assertEquals(paymentSourceId, paymentSource.getId());
-	}
-
-	@Test
-	public void testCustomerSourceRetrieveBitcoinReceiver() throws StripeException {
-		Customer customer = Customer.create(new HashMap<String, Object>());
-		BitcoinReceiver receiver = BitcoinReceiver.create(defaultBitcoinReceiverParams);
-		ExternalAccountCollection customerSources = customer.getSources();
-		Map<String, Object> createParams = new HashMap<String, Object>();
-		createParams.put("source", receiver.getId());
-		customerSources.create(createParams);
-		customerSources = customerSources.all(new HashMap<String, Object>());
-		String paymentSourceId = customerSources.getData().get(0).getId();
-		ExternalAccount paymentSource = customerSources.retrieve(paymentSourceId);
-		assertNotNull(paymentSource);
-		assertEquals(paymentSourceId, paymentSource.getId());
-		assertTrue(paymentSource instanceof BitcoinReceiver);
-		assertTrue(((BitcoinReceiver) paymentSource).getFilled());
-	}
-
-	@Test
-	public void testCustomerCreateWithSource() throws StripeException {
-		HashMap<String, Object> customerCreationParams = new HashMap<String, Object>();
-		customerCreationParams.put("source", defaultSourceParams);
-		Customer customer = Customer.create(customerCreationParams);
-		assertNotNull(customer);
-		assertNotNull(customer.getId());
-		assertNotNull(customer.getSources());
-		assert(customer.getSources().getData().get(0) instanceof Card);
-		assertNotNull(customer.getDefaultSource());
-		ExternalAccount card = customer.getSources().retrieve(customer.getDefaultSource());
-		assertEquals(card.getId(), customer.getDefaultSource());
-	}
-
-	@Test
-	public void testCustomerCreateSourceWithCardHash() throws StripeException {
-		Customer customer = Customer.create(new HashMap<String, Object>());
-		ExternalAccountCollection customerSources = customer.getSources();
-		HashMap<String, Object> createParams = new HashMap<String, Object>();
-		createParams.put("source", defaultSourceParams);
-		ExternalAccount paymentSource = customerSources.create(createParams);
-		assertNotNull(paymentSource);
-		assertNotNull(paymentSource.getId());
-		assert(paymentSource instanceof Card);
-	}
-
-	@Test
-	public void testCustomerCreateSourceWithBitcoinReceiverToken() throws StripeException {
-		Customer customer = Customer.create(new HashMap<String, Object>());
-		ExternalAccountCollection customerSources = customer.getSources();
-		BitcoinReceiver receiver = BitcoinReceiver.create(defaultBitcoinReceiverParams);
-		HashMap<String, Object> createParams = new HashMap<String, Object>();
-		createParams.put("source", receiver.getId());
-		ExternalAccount paymentSource = customerSources.create(createParams);
-		assertNotNull(paymentSource);
-		assertNotNull(paymentSource.getId());
-		assert(paymentSource instanceof BitcoinReceiver);
-	}
-
-	@Test
-	public void testCustomerSourceUpdate() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams);
-		ExternalAccountCollection customerSources = customer.getSources();
-		ExternalAccount paymentSource = customerSources.getData().get(0);
-		assert(paymentSource instanceof Card);
-		Card card = (Card) paymentSource;
-
-		HashMap<String, Object> updateParams = new HashMap<String, Object>();
-		updateParams.put("address_line1", "some address details");
-		Card updatedCard = card.update(updateParams);
-		assertEquals("some address details", updatedCard.getAddressLine1());
-	}
-
-	@Test
-	public void testCustomerSourceDelete() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams);
-		ExternalAccountCollection customerSources = customer.getSources();
-		ExternalAccount paymentSource = customerSources.getData().get(0);
-		paymentSource.delete();
-		HashMap<String, Object> listParams = new HashMap<String, Object>();
-		assertEquals(0, customerSources.all(listParams).getData().size());
-	}
-
-	@Test
-	public void testCustomerCardAddition() throws StripeException {
-		Customer createdCustomer = Customer.create(defaultCustomerParams, supportedRequestOptions);
-		String originalDefaultCard = createdCustomer.getDefaultCard();
-
-		Map<String, Object> creationParams = new HashMap<String, Object>();
-		creationParams.put("card", defaultCardParams);
-		Card addedCard = createdCustomer.createCard(creationParams);
-
-		Token token = Token.create(defaultTokenParams);
-		createdCustomer.createCard(token.getId());
-
-		Customer updatedCustomer = Customer.retrieve(createdCustomer.getId(), supportedRequestOptions);
-		assertEquals((Integer) updatedCustomer.getCards().getData().size(), (Integer) 3);
-		assertEquals(updatedCustomer.getDefaultCard(), originalDefaultCard);
-
-		Map<String, Object> updateParams = new HashMap<String, Object>();
-		updateParams.put("default_card", addedCard.getId());
-		Customer customerAfterDefaultCardUpdate = updatedCustomer.update(updateParams, supportedRequestOptions);
-		assertEquals((Integer) customerAfterDefaultCardUpdate.getCards().getData().size(), (Integer) 3);
-		assertEquals(customerAfterDefaultCardUpdate.getDefaultCard(), addedCard.getId());
-
-		assertEquals(customerAfterDefaultCardUpdate.getCards().retrieve(originalDefaultCard).getId(), originalDefaultCard);
-		assertEquals(customerAfterDefaultCardUpdate.getCards().retrieve(addedCard.getId()).getId(), addedCard.getId());
-	}
-
-	@Test
-	public void testCreateCardThroughCollection() throws StripeException {
-		Customer createdCustomer = Customer.create(defaultCustomerParams, supportedRequestOptions);
-
-		Map<String, Object> creationParams = new HashMap<String, Object>();
-		creationParams.put("card", defaultCardParams);
-		Card addedCard = createdCustomer.getCards().create(creationParams);
-
-		assertEquals(createdCustomer.getId(), addedCard.getCustomer());
-
-		Customer updatedCustomer = Customer.retrieve(createdCustomer.getId(), supportedRequestOptions);
-		assertEquals((Integer) updatedCustomer.getCards().getData().size(), (Integer) 2);
-	}
-
-
-	@Test
-	public void testCustomerCardUpdate() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams, supportedRequestOptions);
-		Card originalCard = customer.getCards().getData().get(0);
-		Map<String, Object> updateParams = new HashMap<String, Object>();
-		updateParams.put("name", "J Bindings Cardholder, Jr.");
-		Card updatedCard = originalCard.update(updateParams);
-		assertEquals(updatedCard.getName(), "J Bindings Cardholder, Jr.");
-	}
-
-	@Test
-	public void testCustomerCardDelete() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams, supportedRequestOptions);
-		Map<String, Object> creationParams = new HashMap<String, Object>();
-		creationParams.put("card", defaultCardParams);
-		customer.createCard(creationParams);
-
-		Card card = customer.getCards().getData().get(0);
-		DeletedCard deletedCard = card.delete();
-		Customer retrievedCustomer = Customer.retrieve(customer.getId(), supportedRequestOptions);
-
-		assertTrue(deletedCard.getDeleted());
-		assertEquals(deletedCard.getId(), card.getId());
-		for(Card retrievedCard : retrievedCustomer.getCards().getData()) {
-				assertFalse("Card was not actually deleted: " + card.getId(), card.getId().equals(retrievedCard.getId()));
-		}
-	}
-
-	@Test
-	public void testCustomerBankAccountAddition() throws StripeException {
-		Customer createdCustomer = Customer.create(defaultCustomerParams, supportedRequestOptions);
-		String originalDefaultCard = createdCustomer.getDefaultCard();
-
-		Map<String, Object> creationParams = new HashMap<String, Object>();
-		creationParams.put("bank_account", defaultBankAccountParams);
-		BankAccount addedBankAccount = createdCustomer.createBankAccount(creationParams);
-
-		Token token = Token.create(defaultTokenParams);
-		createdCustomer.createCard(token.getId());
-
-		Customer updatedCustomer = Customer.retrieve(createdCustomer.getId(), supportedRequestOptions);
-		assertEquals((Integer) updatedCustomer.getSources().getData().size(), (Integer) 3);
-		assertEquals(updatedCustomer.getDefaultCard(), originalDefaultCard);
-
-		Map<String, Object> updateParams = new HashMap<String, Object>();
-		updateParams.put("default_source", addedBankAccount.getId());
-		Customer customerAfterDefaultSourceUpdate = updatedCustomer.update(updateParams, supportedRequestOptions);
-		assertEquals((Integer) customerAfterDefaultSourceUpdate.getSources().getData().size(), (Integer) 3);
-		assertEquals(customerAfterDefaultSourceUpdate.getDefaultSource(), addedBankAccount.getId());
-	}
-
-	@Test
-	public void testCustomerBankAccountDelete() throws StripeException {
-		Customer customer = Customer.create(defaultCustomerParams, supportedRequestOptions);
-
-		Map<String, Object> creationParams = new HashMap<String, Object>();
-		creationParams.put("bank_account", defaultBankAccountParams);
-		BankAccount addedBankAccount = customer.createBankAccount(creationParams);
-
-		DeletedBankAccount deletedBankAccount = addedBankAccount.delete();
-		Customer retrievedCustomer = Customer.retrieve(customer.getId(), supportedRequestOptions);
-
-		assertTrue(deletedBankAccount.getDeleted());
-		assertEquals(deletedBankAccount.getId(), addedBankAccount.getId());
-		for(ExternalAccount retrievedSource : retrievedCustomer.getSources().getData()) {
-			assertFalse("Card was not actually deleted: " + addedBankAccount.getId(),
-					addedBankAccount.getId().equals(retrievedSource.getId()));
-		}
-	}
-
-	@Test
-	public void testCustomerDelete() throws StripeException {
-		Customer createdCustomer = Customer.create(defaultCustomerParams);
-		DeletedCustomer deletedCustomer = createdCustomer.delete();
-		Customer deletedRetrievedCustomer = Customer.retrieve(createdCustomer
-				.getId());
-		assertTrue(deletedCustomer.getDeleted());
-		assertEquals(deletedCustomer.getId(), createdCustomer.getId());
-		assertTrue(deletedRetrievedCustomer.getDeleted());
 	}
 
 	// Plan Tests:
